Rename types to utils
